feat: complete task

+++ src/main/java/ru/clevertec/entity/implementation/RequestEntity.java
@@ -2,10 +2,8 @@ package ru.clevertec.entity.implementation;
 
 import ru.clevertec.entity.AbstractEntity;
 
-import java.math.BigInteger;
-
 public class RequestEntity extends AbstractEntity {
-    public RequestEntity(BigInteger value) {
+    public RequestEntity(Integer value) {
         super(value);
     }
 }
