+++ src/main/java/ru/clevertec/entity/implementation/ResponseEntity.java
@@ -2,10 +2,8 @@ package ru.clevertec.entity.implementation;
 
 import ru.clevertec.entity.AbstractEntity;
 
-import java.math.BigInteger;
-
 public class ResponseEntity extends AbstractEntity {
-    public ResponseEntity(BigInteger value) {
+    public ResponseEntity(Integer value) {
         super(value);
     }
 }
